* move EventHooks to own class files
