+++ library/src/main/java/com/mikepenz/fastadapter/listeners/EventHook.java
@@ -1,0 +1,10 @@
+package com.mikepenz.fastadapter.listeners;
+
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+
+public interface EventHook {
+    @Nullable View onBind(@NonNull RecyclerView.ViewHolder viewHolder);
+}
